[level 0] Title: 두 수의 나눗셈, Time: 0.01 ms, Memory: 60.7 MB -BaekjoonHub

+++ 프로그래머스/lv0/120806. 두 수의 나눗셈/두 수의 나눗셈.kt
@@ -1,3 +1,3 @@
 class Solution {
-    fun solution(num1: Int, num2: Int) = (num1.toFloat() / num2.toFloat() * 1000f).toInt()
+    fun solution(num1: Int, num2: Int) = (num1.toFloat() / num2 * 1000f).toInt()
 }
